Review model cast problem fixed.

+++ lib/features/home/models/review_model.dart
@@ -27,7 +27,7 @@ class ReviewModel extends Equatable {
       userFK: json['userFK'],
       heading: json['heading'],
       review: json['review'],
-      rating: json['rating'],
+      rating: (json['rating'] as int).toDouble(),
     );
   }
